Add more clang-tidy rules (#835)
* Speed up clang-tidy

* Update workflow

* Add new rules

* Add helper for fix

* Formatting

* Fixes from tests

* Remove prepare step from workflow

* Revert change

+++ src/hoi4_world/roles/triggers/always_trigger.h
@@ -23,7 +23,10 @@ class AlwaysTrigger: public Trigger
    [[nodiscard]] bool operator==(const Trigger& rhs) const override { return PolymorphicEquality(*this, rhs); }
    [[nodiscard]] bool operator==(const AlwaysTrigger& rhs) const { return value_ == rhs.value_; }
    [[nodiscard]] bool operator<(const Trigger& rhs) const override { return PolymorphicLess(*this, rhs); }
-   [[nodiscard]] bool operator<(const AlwaysTrigger& rhs) const { return value_ < rhs.value_; }
+   [[nodiscard]] bool operator<(const AlwaysTrigger& rhs) const
+   {
+      return static_cast<int>(value_) < static_cast<int>(rhs.value_);
+   }
 
    [[nodiscard]] std::unique_ptr<Trigger> Copy() const override { return std::make_unique<AlwaysTrigger>(value_); }
    [[nodiscard]] bool IsValid([[maybe_unused]] const Context& context,
